imgui/requirements.txt
glcontext==3.0.0
glfw==2.10.0
imgui-bundle==1.92.4
moderngl==5.12.0
munch==4.0.0
numpy==2.3.4
pillow==12.0.0
pyglm==2.8.2
pyopengl==3.1.10

